Added dbflow local data source
